pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "cryptography>=46.0.3",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google-genai>=1.46.0",
    "html2text>=2025.4.15",
    "imap-tools>=1.11.0",
    "openai>=2.6.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
